[APB-10394] add PersonReference to the IndividualProvidedDetails
